refactor: small nit picks
